src/app/ui/radiobutton/radiobutton.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { RadiobuttonComponent } from './radiobutton.component';

describe('Component: Radiobutton', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [RadiobuttonComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([RadiobuttonComponent],
      (component: RadiobuttonComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(RadiobuttonComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(RadiobuttonComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-radiobutton></app-radiobutton>
  `,
  directives: [RadiobuttonComponent]
})
class RadiobuttonComponentTestController {
}

